Update script based on feedback

+++ scripts/firebase/delete_unused_data.js
@@ -1,25 +1,41 @@
-const prod = false;
-//const prod = true;
-
 const request = require('request-promise');
 
 
+const ENV = process.argv[2];
+let prod = false;
 let baseUrl = '';
-if (prod) {
+if (ENV == 'prod') {
+  prod = true;
   baseUrl = 'https://quillconnect.firebaseio.com';
-} else {
+} else if (ENV == 'staging') {
+  prod = false;
   baseUrl = 'https://quillconnectstaging.firebaseio.com';
+} else {
+  process.stdout.write(
+`This script must be run with a command line flag of either 'prod' or 'staging'.
+In order to actually execute the script against the database include a command line arg of "commit".\n`);
+  process.exit();
 }
 
-const excludeKeys = [
+
+let dryRun = true;
+if (process.argv.includes('commit')) {
+  dryRun = false;
+} else {
+  dryRun = true;
+}
+
+
+const productionKeysInUse = [
   'v2',
   'v3'
 ];
 
-async function fetchSubKeys(ref) {
+
+async function fetchSubKeys(ref, excludeKeys) {
   const options = {
     method: 'GET',
-    uri: `${baseUrl}/${ref || ''}.json?shallow=true`,
+    uri: `${baseUrl}/${ref || ''}.json`,
     qs: {shallow: 'true'},
     json: true
   }
@@ -28,9 +44,11 @@ async function fetchSubKeys(ref) {
       then((keyObj) => {
         const keys = Object.keys(keyObj);
         resolve(keys.filter((key) => !excludeKeys.includes(key)))
-      });
+      }).
+      catch((err) => reject(err));
   })
 }
+
 
 function deleteKey(key) {
   const options = {
@@ -41,14 +59,20 @@ function deleteKey(key) {
     then((resp) => process.stdout.write(`${resp}\n`));
 }
 
-def main() {
-  fetchSubKeys().
+
+function deleteRootLevelKeys(excludeKeys) {
+  fetchSubKeys('', excludeKeys).
     then((keys) => {
       keys.forEach((key) => {
-        process.stdout.write(`${key}\n`);
-        deleteKey(key);
+        if (dryRun) {
+          process.stdout.write(`${key}\n`);
+        } else {
+          deleteKey(key);
+        }
       })
-    })
+    }).
+    catch((err) => process.stdout.write(`${err}\n`));
 }
 
-main();
+
+deleteRootLevelKeys(productionKeysInUse);
